Fix more Quaternion calculations and calibration
Added calibration of the quaternions for sensor fusion IMUs.
Also added pahq and pquat for showing the quaternions from the
AHR and the IMU. And finally, added a flipYaw() virtual function
to flip the sign of yaw after computeAngles() for IMUs like the
ICM20948 that are in LH Android ENU and just quaternion manipulation
isn't getting the yaw to the correct sign in computAngles().